// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-summary-response.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderSummary } from './order-summary';


export interface OrderSummaryResponse { 
    /**
     * offset within the result set, as send in the request
     */
    offset: number;
    /**
     * the maximum number of returned items
     */
    limit: number;
    /**
     * Total number of elements
     */
    total: number;
    summaries?: Array<OrderSummary>;
}