// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-modification-response.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Status } from './status';


export interface OrderModificationResponse { 
    orderIdConsumer?: string;
    status?: Status;
}
export namespace OrderModificationResponse {
}